requirement.txt
asgiref==3.8.1
Django==5.1.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-db-connection-pool==1.2.5
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.9.0
SQLAlchemy==2.0.35
sqlparams==6.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
